+++ src/command-handler.js
@@ -6,6 +6,9 @@ const {downloadResource} = require("./download");
 const {uploadResource} = require("./upload");
 const {getInfo} = require("./info");
 const {runDatabotCommand} = require("./databot");
+const {getAliasesArray} = require("./alias");
+const {getSecretAliasName} = require("./utils");
+
 
 class CommandHandler {
   constructor({tdxConfig, secret, token, timeout}) {
@@ -84,6 +87,34 @@ class CommandHandler {
   async runDatabotCommand({command, id, configJson}) {
     const api = await this.connect();
     return runDatabotCommand({api, command, id, configJson});
+  }
+
+  async runTokenCommand(command) {
+    await this.connect();
+    switch (command) {
+      case "get":
+        return this.accessToken;
+      default:
+        throw Error("Invalid token command");
+    }
+  }
+
+  async getList({alias, type, tdxConfigs, env}) {
+    const aliases = getAliasesArray(tdxConfigs);
+    switch (type) {
+      case "":
+      case "aliases":
+        return {default: alias, aliases};
+      case "secrets":
+        return aliases.reduce(
+          (result, alias) => {
+            result[alias] = env[getSecretAliasName(alias)] || "";
+            return result;
+          }, {}
+        );
+      default:
+        throw Error("Wrong input list type");
+    }
   }
 }