+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.6] - 2020-02-07
+
+### Changed
+
+- Added ```list``` command with ```type=["aliases", "secrets"]```.
+- Added ```--credentials``` option to run commands without signing in.
+- Added ```TDX_CREDENTIALS``` environment variable to run commands without signing in.
+
+### Fixed
+- Script crashed because of unknown request module.
+
 ## [0.2.3] - 2020-02-06
 
 ### Changed